Peak Calculation Depiction
Depicts the MMN data during the post stimulus period of a single participant. Peaks are demarcated with thick dashed lines in each of the MMN (blue) and late MMN (pink) time points. Amplitude ranges used for the calculation of the average MMN amplitude are highlighted with darkened sections within the marked time ranges of interest.