Results of statistical analysis
Depicted are the results of the SRH statistical analysis where the mean of the angry (red) and sad (blue) conditions are plotted with their standard deviation for each visit and time range of interest. Specific average amplitude points for each condition are also represented with diamond (angry) or circle (sad) indicators.